+++ themes/88ds/tpl_rank_list.php
@@ -1,2 +1,0 @@
-<?php if (!defined('__ROOT_DIR__')) exit; ?>
-<?php require __THEME_DIR__ . '/tpl_rank.php'; ?>