+++ themes/88ds/tpl_top.php
@@ -50,7 +50,7 @@
   <?php
   $page_title = '小说排行榜';
   $page_back_url = $site_home_url_raw;
-  require __THEME_DIR__ . '/tpl_page_top.php';
+  ss_render_page_top(['page_title' => $page_title, 'page_back_url' => $page_back_url, 'site_home_url' => $site_home_url_raw]);
   ?>
 
   <div id="content">